Change: Now use dtype and not role

+++ src/main/java/ntnu/idatt2105/madlads/FullstackAPI/controller/UserController.java
@@ -14,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.web.bind.annotation.*;
@@ -52,7 +53,7 @@ public class UserController {
                 returnMap.put("email",foundUser.getEmailAddress());
                 returnMap.put("firstname",foundUser.getFirstName());
                 returnMap.put("lastname",foundUser.getLastName());
-                returnMap.put("role",foundUser.getRole());
+                returnMap.put("role",foundUser.getDtype());
                 returnMap.put("token",generateToken(email));
                 logger.info("Logged in successfully");
                 return new ResponseEntity<>(returnMap, HttpStatus.OK);
@@ -124,10 +125,12 @@ public class UserController {
                 logger.info("trying to register student");
                 String hashedPassword = PasswordHashing.generatePasswordHash(password);
                 QSUser user = new QSUser(firstname, lastname, email, hashedPassword);
-                logger.info(user.getDtype());
                 Student student = userRepository
                         .save(new Student(user));
+                logger.info(student.getDtype());
                 logger.info("Saved new user: " + student.getEmailAddress());
+                QSUser user2 = userRepository.getDistinctByEmailAddress(student.getEmailAddress());
+                logger.info(user2.getDtype());
                 return new ResponseEntity<>(student, HttpStatus.CREATED);
             } catch (Exception e) {
                 logger.info(e.getMessage());
@@ -137,21 +140,28 @@ public class UserController {
             return new ResponseEntity<>(null, HttpStatus.UNPROCESSABLE_ENTITY);
         }
     }
+
     @DeleteMapping("/delete")
     @ResponseStatus(value = HttpStatus.CREATED)
-    public ResponseEntity<QSUser> deleteUser(@RequestParam("email") final String email){
-        QSUser foundUser = userRepository.findByEmailAddress(email);
-        if(foundUser != null){
-            if(email.equals(foundUser.getEmailAddress())){
-                userRepository.delete(foundUser);
-                logger.info("User " + email + " removed");
-                return new ResponseEntity<>(foundUser, HttpStatus.OK);
-            } else {
-                logger.info("User not removed");
+    public ResponseEntity<QSUser> deleteUser(Authentication authentication){
+        if (authentication!=null){
+            if (authentication.isAuthenticated()){
+                String email = authentication.getName();
+                QSUser foundUser = userRepository.findByEmailAddress(authentication.getName());
+                if(foundUser != null){
+                    if(email.equals(foundUser.getEmailAddress())){
+                        userRepository.delete(foundUser);
+                        logger.info("User " + email + " removed");
+                        return new ResponseEntity<>(foundUser, HttpStatus.OK);
+                    } else {
+                        logger.info("User not removed");
+                        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                    }
+                }
+                logger.info("No user with the given email");
                 return new ResponseEntity<>(HttpStatus.NOT_FOUND);
             }
         }
-        logger.info("No user with the given email");
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
